requirements.txt
playwright
beautifulsoup4
langchain-google-genai
langgraph
pydantic
python-dotenv
pytest
pytest-asyncio
google-genai
yt-dlp
